Change mktemp call to make use of new configuration value POT_MKTEMP_SUFFIX. Add $_pname back in to support unit tests.
By default this is set to .XXXXXXX, meaning that it will set 8 random
characters. For unit tests, POT_MKTEMP_SUFFIX can be set to an empty
value or something like .sh, so that it's easy to write
unit tests against those files if necessary.

This means that $TMPDIR is ignored for the time being, which
isn't a big loss, as the rest of pot isn't looking at $TMPDIR
anyway - if pot ever is converted to using $TMPDIR in the
future, this will actually be an easy win/hard to miss.