Garbage collect POSIX shared memory on stop
Addresses parts of #150.

As `posixshmcontrol` doesn't support jails, we simply use
the path of the shared memory segment to determine if
it should be collected, based on if it matches the pot's
base directory.

Anonymous shared memory segments (SHM_ANON) won't show up in
`posixshmcontrol`, but it seems like those are garbage
collected reliably when the process that created them
exits.