Simplify pot start procedure
This aims to address parts of #150.

This changes the jail start procedure from executing the
start command directly to starting a background sleep
process and then using jexec to run the actual command.

Also, this removes modifying the running jail to be
non-persistent and instead just send nopersist in
parameters.

This has a couple of advantages:
- No need for the bg_start process with its wicked timing
- Processes have an easier cleaning up after themselves
- Code slightly easier to understand.

Downsides:
- Pot requires /bin/sh and /bin/sleep
- This requires quite some testing

Right now, the poststart script (just like the
prestart script and the pre-/poststart scripts) runs in
the foreground (before it was part of the background
process). I think this is actually better, but needs
to be checked.