Prevent stopping non-persistent jails twice
This aims to address #152.

The problem at hand isn't really to prevent `pot stop` from
running in parallel (this could be handled on a global level),
but to prevent running "_js_stop" in `pot start`, after a
non-persistent jail was stopped.

The solution is very simple, as the use case this is about is:
1. pot stop is called
2. jail -r is called
3. pot start continues and calls "_js_stop"

By removing a file (and touching it when running "_js_stop")
this can be prevented without having to use real lock files.

Note: Given the number of files pot creates, it might make sense
to start using a subdirectory called _/tmp/pot_.