Remove leftover pot-stopped files (#179)
This only works on non-persistent pots or in case pot start failed
(and stop still happened, which is not a likely scenario). In the
absence of a more clever mechanism of stop knowing that start
is still running, this seems like a nice and simple trade-off - leave
pot stopped files behind with persistent jails (which usually use
names that allow files to be overwritten on next start), but remove them
when running from nomad (which creates random pot names and starts/stops
at a higher frequency when operating the cluster).