Allow specifying pf_rules file to patch on init
When altering more complex setups automatically, pot's
way of patching pf.conf is a bit too simplistic.

By adding this flag, the user has multiple choices of modifying
pf.conf in a controlled way:

1. Write to a different file that's included in pf.conf
   using its include keyword.
2. Write to a different file that's parsed by additional
   tooling to assemble a pf.conf (e.g., in automation).
3. In setups where pots modifications serve no real purpose,
   running `pot init -f ''` can be used to skip touching
   pf.conf completely.

Help text intentionally shows the default to be determined
by a command, not the result of it.

This is the first patch of a series, with more complex ones
to follow, which serve the ultimate purpose of making pot's
networking more flexible/customizable using hooks, so it
can be integrated into different network environments (as
one size doesn't fit all).