Inherit encryption property from parent filesystem
This makes sure that pot running on an encrypted parent fs will
also be encrypted by inheriting the property.

This is implemented in a way that should be backwards compatible
on ZFS versions not supporting the encryption property.

New common function _is_zfs_property_supported implemented so
that it is multi-purpose.