Label and group interfaces created by pot
When running a large number of pots, it gets hard to easily figure out which
epair interface was created for which pot, especially when debugging and if
the jail is not running anymore.

This change adds two attributes to all interfaces created by pot:

1. A description, which is the pot name in case pots/epair interfaces and
   the bridge's purpose in case of bridges.
2. The group "pot"

With the change, it's easy to find interfaces belonging to pot

    $ ifconfig -g pot
    bridge0
    epair2a
    epair3a
    epair4a
    epair5a
    epair6a
    epair7a
    epair8a
    epair9a

and also easier to figure out what their purpose is

    $ ifconfig bridge0 | grep -B1 descr
    bridge0: flags=8843<UP,BROADCAST,RUNNING,SIMPLEX,MULTICAST> metric 0...
        description: pot public bridge

    $ ifconfig epair2a
    epair2a: flags=8943<UP,BROADCAST,RUNNING,PROMISC,SIMPLEX,MULTICAST> ...
            description: my-pot_96958fcf_1401fde2-d6e9-f12e-9c19-08baf0195561
            options=8<VLAN_MTU>
            ether 02:5d:2b:d4:4e:0a
            groups: epair pot
            media: Ethernet 10Gbase-T (10Gbase-T <full-duplex>)
            status: active
            nd6 options=29<PERFORMNUD,IFDISABLED,AUTO_LINKLOCAL>

    $ ifconfig -g pot | xargs -n1 ifconfig | grep -B1 my-pot
    epair2a: flags=8943<UP,BROADCAST,RUNNING,PROMISC,SIMPLEX,MULTICAST> ...
            description: my-pot_96958fcf_1401fde2-d6e9-f12e-9c19-08baf0195561

The information contained in descriptions/groups could also be taken advantage
of when developing a more robust `pot stop` procedure.