Store tinirc.pid on start, truncate tinirc on start
Storing the pid (which will become the pid of the process tinirc
starts) makes writing stop scripts easier.

Picked /tmp as the location, as /var/run might not be available in
slim no-rc images.

Fixed a bug where an existing copy of tinirc would be appended
to, meaning that none of the pot's settings would actually be
applied.