import/prepare: Setting a default signature public key
This allows to set a global default key to check against, which can
still be overridden by `-C pubkey` in import/prepare.

This is especially useful when using the nomad-pot-plugin, as it
allows to limit the pot images to run on the cluster to those
which were signed by a trusted pot provider.